knn/requirements.txt
numpy-1.13.3
pandas-0.20.3
scikit-learn-0.19.0
sklearn
